requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
anthropic>=0.49.0
langchain==0.3.9
langchain-anthropic==0.3.3
langgraph==0.2.55
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
jinja2==3.1.4
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.3
aiohttp==3.11.9
authlib==1.3.2
itsdangerous==2.2.0
starlette==0.41.3